requirements.txt
torch>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
pyyaml>=6.0
streamlit>=1.32.0
plotly>=5.0.0
